% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_DNAm.r
\name{estlambda}
\alias{estlambda}
\title{Calculate lambda from genome-wide association results}
\usage{
estlambda(pvals)
}
\arguments{
\item{pvals}{a vector of p-values to estimate lambda from}
}
\value{
the lambda value
}
\description{
Calculate lambda from genome-wide association results
}
\author{
E Hannon
}
