% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_DNAm.r
\name{getScanDate}
\alias{getScanDate}
\title{Extract scan date from idat file}
\usage{
getScanDate(idat)
}
\arguments{
\item{idat}{file path to a single idat file from either colour channel for a sample e.g. 205096370026_R02C01_Red.idat}
}
\value{
date idat file was scanned by
}
\description{
Extract scan date from idat file
}
\author{
E Hannon
}
