% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_DNAm.r
\name{reformatBetas}
\alias{reformatBetas}
\title{Reformat beta matrix so rownames ordered to match a specificed list}
\usage{
reformatBetas(bMat, pList)
}
\arguments{
\item{bMat}{matrix of beta values}

\item{pList}{a vector of rownames in the required order}
}
\value{
a reordered matrix of beta values
}
\description{
Reformat beta matrix so rownames ordered to match a specificed list
}
\author{
E Hannon
}
